move missed classes to export package and move data export classes to export package
