+++ src/main/java/de/tum/in/www1/artemis/service/export/DataExportCommunicationDataService.java
@@ -1,8 +1,8 @@
-package de.tum.in.www1.artemis.service.dataexport;
+package de.tum.in.www1.artemis.service.export;
 
-import static de.tum.in.www1.artemis.service.dataexport.DataExportExerciseCreationService.CSV_FILE_EXTENSION;
-import static de.tum.in.www1.artemis.service.dataexport.DataExportUtil.createDirectoryIfNotExistent;
-import static de.tum.in.www1.artemis.service.dataexport.DataExportUtil.retrieveCourseDirPath;
+import static de.tum.in.www1.artemis.service.export.DataExportExerciseCreationService.CSV_FILE_EXTENSION;
+import static de.tum.in.www1.artemis.service.export.DataExportUtil.createDirectoryIfNotExistent;
+import static de.tum.in.www1.artemis.service.export.DataExportUtil.retrieveCourseDirPath;
 
 import java.io.IOException;
 import java.nio.file.Files;
